+++ src/rapid_router/models/__init__.py
@@ -4,4 +4,5 @@ Created on 03/04/2024 at 11:11:08(+01:00).
 """
 
 from .level import Level
+from .top_score import Attempt
 from .user import User
