+++ src/rapid_router/models/top_score.py
@@ -1,0 +1,8 @@
+"""
+© Ocado Group
+Created on 30/10/2025 at 17:08:12(+00:00).
+"""
+
+# pylint: disable-next=unused-import
+# TODO: rename this model to TopScore
+from game.models import Attempt  # type: ignore[import-untyped]
